+++ src/backend/core/tests/documents/test_api_documents_attachment_upload.py
@@ -67,9 +67,11 @@ def test_api_documents_attachment_upload_anonymous_success():
     file_path = response.json()["file"]
     match = pattern.search(file_path)
     file_id = match.group(1)
-
     # Validate that file_id is a valid UUID
     uuid.UUID(file_id)
+
+    document.refresh_from_db()
+    assert document.attachments == [f"{document.id!s}/attachments/{file_id!s}.png"]
 
     # Now, check the metadata of the uploaded file
     key = file_path.replace("/media", "")
@@ -112,6 +114,9 @@ def test_api_documents_attachment_upload_authenticated_forbidden(reach, role):
         "detail": "You do not have permission to perform this action."
     }
 
+    document.refresh_from_db()
+    assert document.attachments == []
+
 
 @pytest.mark.parametrize(
     "reach, role",
@@ -122,8 +127,8 @@ def test_api_documents_attachment_upload_authenticated_forbidden(reach, role):
 )
 def test_api_documents_attachment_upload_authenticated_success(reach, role):
     """
-    Autenticated who are not related to a document should be able to upload a file
-    if the link reach and role permit it.
+    Autenticated users who are not related to a document should be able to upload
+    a file when the link reach and role permit it.
     """
     user = factories.UserFactory()
 
@@ -144,6 +149,9 @@ def test_api_documents_attachment_upload_authenticated_success(reach, role):
 
     # Validate that file_id is a valid UUID
     uuid.UUID(file_id)
+
+    document.refresh_from_db()
+    assert document.attachments == [f"{document.id!s}/attachments/{file_id!s}.png"]
 
 
 @pytest.mark.parametrize("via", VIA)
@@ -174,6 +182,9 @@ def test_api_documents_attachment_upload_reader(via, mock_user_teams):
     assert response.json() == {
         "detail": "You do not have permission to perform this action."
     }
+
+    document.refresh_from_db()
+    assert document.attachments == []
 
 
 @pytest.mark.parametrize("role", ["editor", "administrator", "owner"])
@@ -211,6 +222,9 @@ def test_api_documents_attachment_upload_success(via, role, mock_user_teams):
     # Validate that file_id is a valid UUID
     uuid.UUID(file_id)
 
+    document.refresh_from_db()
+    assert document.attachments == [f"{document.id!s}/attachments/{file_id!s}.png"]
+
     # Now, check the metadata of the uploaded file
     key = file_path.replace("/media", "")
     file_head = default_storage.connection.meta.client.head_object(
@@ -236,6 +250,9 @@ def test_api_documents_attachment_upload_invalid(client):
     assert response.status_code == 400
     assert response.json() == {"file": ["No file was submitted."]}
 
+    document.refresh_from_db()
+    assert document.attachments == []
+
 
 def test_api_documents_attachment_upload_size_limit_exceeded(settings):
     """The uploaded file should not exceeed the maximum size in settings."""
@@ -257,6 +274,9 @@ def test_api_documents_attachment_upload_size_limit_exceeded(settings):
 
     assert response.status_code == 400
     assert response.json() == {"file": ["File size exceeds the maximum limit of 1 MB."]}
+
+    document.refresh_from_db()
+    assert document.attachments == []
 
 
 @pytest.mark.parametrize(
@@ -293,6 +313,11 @@ def test_api_documents_attachment_upload_fix_extension(
     match = pattern.search(file_path)
     file_id = match.group(1)
 
+    document.refresh_from_db()
+    assert document.attachments == [
+        f"{document.id!s}/attachments/{file_id!s}.{extension:s}"
+    ]
+
     assert "-unsafe" in file_id
     # Validate that file_id is a valid UUID
     file_id = file_id.replace("-unsafe", "")
@@ -323,6 +348,9 @@ def test_api_documents_attachment_upload_empty_file():
     assert response.status_code == 400
     assert response.json() == {"file": ["The submitted file is empty."]}
 
+    document.refresh_from_db()
+    assert document.attachments == []
+
 
 def test_api_documents_attachment_upload_unsafe():
     """A file with an unsafe mime type should be tagged as such."""
@@ -344,6 +372,9 @@ def test_api_documents_attachment_upload_unsafe():
     pattern = re.compile(rf"^/media/{document.id!s}/attachments/(.*)\.exe")
     match = pattern.search(file_path)
     file_id = match.group(1)
+
+    document.refresh_from_db()
+    assert document.attachments == [f"{document.id!s}/attachments/{file_id!s}.exe"]
 
     assert "-unsafe" in file_id
     # Validate that file_id is a valid UUID